+++ src/compas_view2/views/view.py
@@ -19,17 +19,17 @@ class View(QtWidgets.QOpenGLWidget):
         The view configuration.
     """
 
-    VIEW_PORTS = {"FRONT": 1, "RIGHT": 2, "TOP": 3, "PERSPECTIVE": 4}
+    VIEW_PORTS = {"front": 1, "right": 2, "top": 3, "perspective": 4}
 
     def __init__(self, app, view_config):
         super().__init__()
         self.setFocusPolicy(QtCore.Qt.StrongFocus)
         self._opacity = 1.0
-        self._current = self.VIEW_PORTS[view_config["view_port"]]
+        self._current = self.VIEW_PORTS[view_config["viewport"]]
         self.shader_model = None
         self.app = app
         self.color = view_config["background_color"]
-        self.mode = view_config["view_mode"]
+        self.mode = view_config["viewmode"]
         self.selection_color = view_config["selection_color"]
         self.show_grid = view_config["show_grid"]
         self.camera = Camera(self, **view_config["camera"])